+++ src/easy_chat.c
@@ -5004,7 +5004,7 @@ static void HideStartSelectButtons(void)
 
 static void TryAddInterviewObjectEvents(void)
 {
-    int graphicsId;
+    u16 graphicsId;
     u8 spriteId;
 
     switch (GetDisplayedPersonType())
